+++ ios/TwilioVoiceReactNative.h
@@ -43,8 +43,6 @@ FOUNDATION_EXPORT NSString * const kTwilioVoiceReactNativeEventKeyCancelledCallI
 
 + (TVODefaultAudioDevice *)twilioAudioDevice;
 
-@property(nonatomic, strong) NSString *incomingCallContactHandleTemplate;
-
 @end
 
 @interface TwilioVoiceReactNative (EventEmitter)